# cmd/api/docs/swagger.yaml
definitions:
  go-clean-arch_internal_adapter_delivery_http_dto.LoginBody:
    properties:
      email:
        type: string
      password:
        maxLength: 25
        minLength: 5
        type: string
    required:
    - email
    - password
    type: object
  go-clean-arch_internal_adapter_delivery_http_dto.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status:
        type: integer
    type: object
  go-clean-arch_internal_adapter_delivery_http_dto.UserSignUp:
    properties:
      email:
        type: string
      password:
        maxLength: 25
        minLength: 5
        type: string
      username:
        maxLength: 15
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  internal_adapter_delivery_http_handler.Response:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      description: API for user to login
      operationId: UserLogin
      parameters:
      - description: Input Fields
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.LoginBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
      summary: Login (User)
      tags:
      - Auth
  /api/auth/signup:
    post:
      description: API for user to register a new account
      operationId: UserSignUp
      parameters:
      - description: Input Fields
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.UserSignUp'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
      summary: Signup (User)
      tags:
      - Auth
  /api/user/me:
    get:
      description: API for user to get their own profile
      operationId: UserMe
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_delivery_http_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-clean-arch_internal_adapter_delivery_http_dto.Response'
      security:
      - Bearer: []
      summary: Get User Profile
      tags:
      - User
swagger: "2.0"